+++ src/engines/index.ts
@@ -1,7 +1,7 @@
 import * as configuration from "../configuration";
 import * as dot from "./dot";
 
-function getCurrentEngine(): (source: string) => Promise<string> {
+function getCurrentEngine(): (source: string, cancel: Promise<void>) => Promise<string> {
     const engineName = configuration.getNullableConfiguration("engine", "dot");
 
     switch (engineName) {
@@ -14,8 +14,8 @@ function getCurrentEngine(): (source: string) => Promise<string> {
 
 let currentEngine = getCurrentEngine();
 
-export function getEngine(): (source: string) => Promise<string> {
-    return (source) => currentEngine(source); // Do not use `return currentEngine;`.
+export function getEngine(): (source: string, cancel: Promise<void>) => Promise<string> {
+    return (source, cancel) => currentEngine(source, cancel); // Do not use `return currentEngine;`.
 }
 
 export function updateConfiguration(): void {